[Nikhil]refactor class defined

+++ Automapping/EmployeeBlrBranch.cs
@@ -6,7 +6,7 @@
     public  class EmployeeBlrBranch
     {
         public string FullName { get; set; }
-        public int Contact { get; set; }
+        public long Contact { get; set; }
         public string Address { get; set; }
         public int EmpId { get; set; }
         public string Department { get; set; }
